requirements.txt
flask
pyjwt
python-dotenv
mysql-connector-python